requirements.txt
requests~=2.32.3
pandas~=2.2.3
tweepy~=4.14.0
PyYAML~=6.0.2